@@ -1,6 +1,6 @@
 {
   "name": "mkdocs-material",
-  "version": "6.0.2",
+  "version": "6.2.6",
   "description": "A Material Design theme for MkDocs",
   "keywords": [
     "mkdocs",
@@ -34,66 +34,66 @@
   "dependencies": {
     "clipboard": "^2.0.6",
     "escape-html": "^1.0.3",
-    "focus-visible": "^5.1.0",
+    "focus-visible": "^5.2.0",
     "lunr": "^2.3.9",
     "lunr-languages": "^1.4.0",
-    "resize-observer-polyfill": "^1.5.1",
-    "rxjs": "^7.0.0-beta.7"
+    "rxjs": "^7.0.0-beta.10"
   },
   "devDependencies": {
-    "@fortawesome/fontawesome-free": "^5.15.0",
-    "@mdi/svg": "^5.6.55",
-    "@primer/octicons": "^11.0.0",
+    "@fortawesome/fontawesome-free": "^5.15.2",
+    "@mdi/svg": "^5.9.55",
+    "@primer/octicons": "^11.2.0",
     "@types/clipboard": "^2.0.1",
-    "@types/copy-webpack-plugin": "^6.0.0",
+    "@types/copy-webpack-plugin": "^6.4.0",
     "@types/escape-html": "1.0.0",
-    "@types/event-hooks-webpack-plugin": "^2.0.0",
+    "@types/event-hooks-webpack-plugin": "^2.2.0",
     "@types/html-minifier": "^4.0.0",
     "@types/lunr": "^2.3.3",
-    "@types/mini-css-extract-plugin": "^0.9.1",
-    "@types/node": "^14.11.2",
-    "@types/ramda": "^0.27.20",
-    "@types/resize-observer-browser": "^0.1.3",
-    "@types/webpack": "^4.41.22",
+    "@types/mini-css-extract-plugin": "^1.2.2",
+    "@types/node": "^14.14.22",
+    "@types/ramda": "^0.27.36",
+    "@types/resize-observer-browser": "^0.1.5",
+    "@types/webpack": "^4.41.26",
     "@types/webpack-assets-manifest": "^3.0.1",
-    "autoprefixer": "9.8.6",
-    "copy-webpack-plugin": "^6.2.0",
-    "css-loader": "^4.3.0",
-    "event-hooks-webpack-plugin": "^2.1.7",
-    "expose-loader": "^1.0.0",
+    "autoprefixer": "10.2.3",
+    "copy-webpack-plugin": "^6.4.1",
+    "css-loader": "^5.0.1",
+    "event-hooks-webpack-plugin": "^2.2.0",
+    "expose-loader": "^1.0.3",
     "github-types": "^1.0.0",
     "gitlab": "^14.2.2",
     "html-minifier": "^4.0.0",
-    "ignore-emit-webpack-plugin": "^2.0.3",
+    "ignore-emit-webpack-plugin": "2.0.3",
     "imagemin-webpack-plugin": "^2.4.2",
     "material-design-color": "^2.3.2",
     "material-shadows": "^3.0.1",
-    "mini-css-extract-plugin": "^0.11.3",
-    "postcss-inline-svg": "^4.1.0",
-    "postcss-loader": "^4.0.3",
+    "mini-css-extract-plugin": "^1.3.4",
+    "postcss": "^8.2.4",
+    "postcss-inline-svg": "^5.0.0",
+    "postcss-loader": "^4.2.0",
     "postcss-svgo": "^4.0.2",
-    "preact": "^10.5.3",
+    "preact": "^10.5.11",
     "ramda": "^0.27.1",
     "rimraf": "^3.0.2",
-    "sass": "^1.26.11",
-    "sass-loader": "^10.0.2",
-    "stylelint": "^13.7.2",
+    "sass": "^1.32.5",
+    "sass-loader": "^10.1.1",
+    "stylelint": "^13.9.0",
     "stylelint-config-rational-order": "^0.1.2",
     "stylelint-config-standard": "^20.0.0",
     "stylelint-order": "^4.1.0",
     "stylelint-scss": "^3.18.0",
-    "ts-loader": "^8.0.4",
-    "ts-node": "^9.0.0",
+    "ts-loader": "^8.0.14",
+    "ts-node": "^9.1.1",
     "tsconfig-paths-webpack-plugin": "^3.3.0",
-    "tslib": "^2.0.1",
+    "tslib": "^2.1.0",
     "tslint": "^6.1.3",
     "tslint-sonarts": "^1.9.0",
-    "typescript": "^3.9.7",
+    "typescript": "^4.1.3",
     "webpack": "^4.44.2",
-    "webpack-assets-manifest": "^3.1.1",
-    "webpack-cli": "^3.3.12"
+    "webpack-assets-manifest": "3.1.1",
+    "webpack-cli": "^4.4.0"
   },
   "engines": {
-    "node": ">= 10"
+    "node": ">= 12"
   }
 }
