@@ -52,13 +52,29 @@ export function resetSidebarOffset(
 /**
  * Set sidebar height
  *
+ * This function doesn't set the height of the actual sidebar, but of its first
+ * child – the `.md-sidebar__scrollwrap` element in order to mitigiate jittery
+ * sidebars when the footer is scrolled into view. At some point we switched
+ * from `absolute` / `fixed` positioning to `sticky` positioning, which greatly
+ * reduced jitter in some browsers (respectively Firefox and Safari) when
+ * scrolling from the top. However, top-aligned sticky positioning means that
+ * the sidebar snaps to the bottom when the end of the container is reached.
+ * This is what leads to the mentioned jitter, as the sidebar's height may be
+ * updated to slowly.
+ *
+ * By setting the height of the sidebar to zero (while preserving `padding`),
+ * and the height on its first element, this behaviour can be mitigiated. We
+ * must assume that the top- and bottom offset (`padding`) are equal, as the
+ * `offsetBottom` value is `undefined`.
+ *
  * @param el - Sidebar element
  * @param value - Sidebar height
  */
 export function setSidebarHeight(
   el: HTMLElement, value: number
 ): void {
-  el.style.height = `${value}px`
+  const scrollwrap = el.firstElementChild as HTMLElement
+  scrollwrap.style.height = `${value - 2 * scrollwrap.offsetTop}px`
 }
 
 /**
@@ -69,5 +85,6 @@ export function setSidebarHeight(
 export function resetSidebarHeight(
   el: HTMLElement
 ): void {
-  el.style.height = ""
+  const scrollwrap = el.firstElementChild as HTMLElement
+  scrollwrap.style.height = ""
 }
